src/requirements.txt
# Azure version dependencies
semantic-kernel>=1.0.0
openai>=1.0.0
azure-identity>=1.15.0
python-dotenv>=1.0.0

# Local Claude version dependencies
anthropic>=0.40.0

# Common
asyncio
